Updated package name to view. Added ComputerTest
